services: rename streamStatus to streamActivity
